refactor: move `ListDataset` to utils
