+++ src/vidore_benchmark/retrievers/colqwen_retriever.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from typing import ClassVar, List, Optional, TypeVar, cast
+from typing import ClassVar, List, Optional, cast
 
 import torch
 from colpali_engine.models import ColQwen2, ColQwen2Processor
@@ -8,25 +8,14 @@ from colpali_engine.utils.torch_utils import get_torch_device
 from dotenv import load_dotenv
 from loguru import logger
 from PIL import Image
-from torch.utils.data import DataLoader, Dataset
+from torch.utils.data import DataLoader
 from tqdm import tqdm
 
 from vidore_benchmark.retrievers.utils.register_retriever import register_vision_retriever
 from vidore_benchmark.retrievers.vision_retriever import VisionRetriever
+from vidore_benchmark.utils.data_utils import ListDataset
 
-T = TypeVar("T")
 load_dotenv(override=True)
-
-
-class ListDataset(Dataset[T]):
-    def __init__(self, elements: List[T]):
-        self.elements = elements
-
-    def __len__(self) -> int:
-        return len(self.elements)
-
-    def __getitem__(self, idx: int) -> T:
-        return self.elements[idx]
 
 
 @register_vision_retriever("colqwen2")
