+++ src/vidore_benchmark/utils/data_utils.py
@@ -1,0 +1,16 @@
+from typing import List, TypeVar
+
+from torch.utils.data import Dataset
+
+T = TypeVar("T")
+
+
+class ListDataset(Dataset[T]):
+    def __init__(self, elements: List[T]):
+        self.elements = elements
+
+    def __len__(self) -> int:
+        return len(self.elements)
+
+    def __getitem__(self, idx: int) -> T:
+        return self.elements[idx]
